fix warning

+++ src/bartz/mcmcstep.py
@@ -508,7 +508,7 @@ def grow_move(X, var_tree, split_tree, max_split, p_nonterminal, sigma2, resid, 
     leaf_to_grow, num_growable, num_prunable = choose_leaf(split_tree, key1)
     
     var, num_available_var = choose_variable(var_tree, max_split, leaf_to_grow, key2)
-    var_tree = var_tree.at[leaf_to_grow].set(var)
+    var_tree = var_tree.at[leaf_to_grow].set(var.astype(var_tree.dtype))
     
     split, num_available_split = choose_split(var_tree, split_tree, max_split, leaf_to_grow, key3)
     split_tree = split_tree.at[leaf_to_grow].set(split)
